Force views.serve() to return a UTF-8 bytestring.
This is backwards-compatible, as it changes none of the APIs end users
work with, but ensures that the returned XML is actually UTF-8 bytes,
since the serialized Policy does not declare an encoding in its prolog
and thus *must* be UTF-8.

Also included are slight tweaks to the tests to ensure the behavior is
correct.