Clean up str/bytes handling in Policy serialization.
The basic issue here is that __str__() cannot do everything that's
needed, in terms of correctly serializing to UTF-8-encoded XML,
because on Python 3 it must return a Unicode string.

The solution implemented here is to keep __str__() as a debugging aid,
and add a new method to Policy -- serialize() -- which can and does
return a bytes object encoded in UTF-8.

This is backwards compatible, since __str__() continues to function as
in previous versions, but allows anything which will serve a Policy or
write it to a file to have a shorter one-line way to get it in the
desired (i.e., UTF-8-encoded) format.